Change AbstractPriceStrategy & PriceStrategy
Rosco and I decided to change this interface so we don't return the
inversed price but the tokenprice (USD). This is WIP and will break all
pricing strategies atm. I've started working on UniswapV2PriceStrategy,
just to test it out.

The idea is that we will add an extra RPC call that returns the
decimals. We will use the viem PublicClient cacheTime to make up for
this.